+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Williams](https://togaware.com/Graham.Williams.html).
 
 ## 0.2 Basic Functionality
 
++ ADD TASK button for KANBAN page [0.1.51 20260612 gjw]
 + Sorting added to KANBAN [0.1.50 20260612 gjw]
 + Update solid_auth to fix token refresh bug [0.1.49 20260612 gjw]
 + Add SAVE to PDF VIEW in BACKUP [0.1.48 20260612 gjw]